test(testgen): split BehavioralTest into per-feature suites
BehavioralTest.scala had grown to 1604 lines (the largest code file) with
57 tests accreted across feature areas (M5.9, #152, #155, status-restriction).
Split it along its section boundaries into focused suites sharing a fixture
trait:

- BehavioralTestSupport (trait extends CatsEffectSuite): the shared fixtures
  (asSynthesized, loadProfiled[FromSpec], profileSource, sensitiveInputSpec,
  cardinalitySpec), protected so subclasses inherit them.
- BehavioralTest: core single-shot ensures/negative/invariant/sensitive tests.
- BehavioralTransitionTest: M5.9 TransitionDecl-aware property tests.
- BehavioralGuardedTransitionTest: #152 guarded-transition seed-dict recognizer.
- BehavioralViaOpTest: #155 via-ops with body/query inputs.
- BehavioralStatusRestrictionTest: status-restriction negative tests.

Pure test relocation, no assertions changed. 57 behavioral tests preserved
(now 5 discovered suites); full testgen suite 287/287 green; scalafmt +
scalafix clean; compiles under -Werror.
